pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "Fast-Permissions"
dynamic = ["version"]
authors = [{name="Cody M Sommer", email="bassmastacod@gmail.com"}]
description = "Add robust authentication to your FastAPI endpoints"
keywords = ["authentication", "authorization", "fastapi", "permissions", "security", "access"]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["fastapi", "daomodel", "pyjwt", "python-multipart", "jinja2", "bcrypt"]
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed"
]

[project.urls]
Repository = "https://github.com/BassMastaCod/Fast-Permissions.git"
Issues = "https://github.com/BassMastaCod/Fast-Permissions/issues"

[tool.pdm.version]
source = "scm"

[tool.pytest.ini_options]
pythonpath = "fast_permissions"
addopts = [
    "--import-mode=importlib",
]

[project.optional-dependencies]
pwa = ["fastpwa[controller]"]
